Adjusting after PR. Refactoring. Minor changes.
