+++ src/main/java/edu/kis/powp/jobs2d/drivers/adapter/DrawLineAdapter.java
@@ -1,14 +1,16 @@
-package edu.kis.powp.jobs2d.drivers;
+package edu.kis.powp.jobs2d.drivers.adapter;
 
 import edu.kis.legacy.drawer.shape.ILine;
 import edu.kis.powp.jobs2d.AbstractDriver;
 import edu.kis.legacy.drawer.shape.LineFactory;
-import edu.kis.powp.jobs2d.features.DrawerFeature;
 
-public class DrawLineManager extends AbstractDriver {
+
+import static edu.kis.powp.jobs2d.features.DrawerFeature.getDrawerController;
+
+public class DrawLineAdapter extends AbstractDriver {
     private ILine line;
 
-    public DrawLineManager(int x, int y, ILine line){
+    public DrawLineAdapter(int x, int y, ILine line){
         super(x,y);
         this.line=line;
     }
@@ -18,7 +20,7 @@ public class DrawLineManager extends AbstractDriver {
         line.setStartCoordinates(this.getX(), this.getY());
         line.setEndCoordinates(x, y);
         this.setPosition(x, y);
-        DrawerFeature.getDrawerController().drawLine(line);
+        getDrawerController().drawLine(line);
     }
     @Override
     public String toString() {
