+++ src/main/java/edu/kis/powp/jobs2d/drivers/adapter/LineDrawerAdapter.java
@@ -6,7 +6,7 @@ import edu.kis.powp.jobs2d.Job2dDriver;
 import edu.kis.powp.jobs2d.features.DrawerFeature;
 
 
-public class LineDrawerAdapter extends DrawPanelController implements Job2dDriver {
+public class LineDrawerAdapter implements Job2dDriver {
 
     private int startX = 0;
     private int startY = 0;
