+++ src/test/java/edu/kis/powp/jobs2d/events/SelectTestFigureOptionListener.java
@@ -9,14 +9,14 @@ import edu.kis.powp.jobs2d.magicpresets.FiguresJoe;
 public class SelectTestFigureOptionListener implements ActionListener {
 
 	private DriverManager driverManager;
-
 	public SelectTestFigureOptionListener(DriverManager driverManager) {
 		this.driverManager = driverManager;
 	}
 
 	@Override
 	public void actionPerformed(ActionEvent e) {
-		if(e.getActionCommand().equals("Figure Joe 1"))
+		String test = "Figure Joe 1";
+		if(e.getActionCommand().equals(test))
 		{
 			FiguresJoe.figureScript1(driverManager.getCurrentDriver());
